requirements.txt
spotipy
numpy
flask
python-dotenv